code clean

+++ lib/Common/heat_state.h
@@ -12,9 +12,9 @@ struct HeatState {
   uint32_t lastWaterPumpedTimestamp;        // mills
   float heatBalancePool;                    // Unit is Joule == Watt*Second
   uint32_t lastPidOutputTimestamp;          // mills
-  double pidOutput;                         // pid out, how many milis boiler on in 1 second, no need to reset when brew on-off
+  float pidOutput;                          // pid out, how many milis boiler on in 1 second, no need to reset when brew on-off
   float lastThermoCompensateHeat;           // Thermo Compensation heat caculate, Unit is Joule
-  float lastThermoHeaterConsumed;             // Unit is Joule
+  float lastThermoHeaterConsumed;           // Unit is Joule
   bool isBoilerOperatorTC;                  // Who operate the boiler, true is Thermo Compensation, false is PID or other.
 
 };
